feat: SHR and SHL instructions implementation.

+++ src/instructions/shl/shl.h
@@ -3,6 +3,6 @@
 #ifndef SHL_ISTR
 #define SHL_ISTR
 
-void SHL(CPUContext *cpuCtxPtr, uint8_t rd, uint8_t rm, uint16_t immediate);
+void SHL(CPUContext *cpuCtxPtr, uint8_t rd, uint8_t rm, uint8_t immediate);
 
 #endif
