+++ src/instructions/shr/shr.h
@@ -3,6 +3,6 @@
 #ifndef SHR_ISTR
 #define SHR_ISTR
 
-void SHR(CPUContext *cpuCtxPtr, uint8_t rd, uint8_t rm, uint16_t immediate);
+void SHR(CPUContext *cpuCtxPtr, uint8_t rd, uint8_t rm, uint8_t immediate);
 
 #endif
